pyproject.toml
[project]
name = "project_artemis"
version = "0.1.0"
description = "AI-powered CLI tool for generating content using Google's Gemini API. Lightweight, secure, and extensible."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
